Changed the folder of integration tests to the default value
Our first iteration on the integration framework used a specific folder
under each project to classify tests as unit test or integration tests.
It works fine, but IDEs require some configuration to recognized the 
folder. We have change that to use a more standard approach on which
tests are, by default, unit tests, but can be annotated with
@Category(IntegrationTest.class*) to indicate they are integration
tests and therefore must be ignored by surefire pluging and executed
with failsafe plugin

\* IntegrationTest.class is on the core module
