+++ engine/core/src/main/java/com/torodb/core/test/IntegrationTest.java
@@ -1,0 +1,30 @@
+/*
+ * ToroDB
+ * Copyright © 2014 8Kdata Technology (www.8kdata.com)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.torodb.core.test;
+
+/**
+ * This interface is used to mark tests that must be running with the integration framework (
+ * usually Maven Failsafe).
+ *
+ * To do that, just annotate the test with
+ * {@literal @}{@link org.junit.experimental.categories.Category}(IntegrationTest.class)
+ */
+public interface IntegrationTest {
+
+}
